# perf(reports) use lock-free lua-resty-counter in reports