--- a/kong-1.4.2-0.rockspec
+++ b/kong-1.4.2-0.rockspec
@@ -36,7 +36,7 @@ dependencies = {
   "lua-resty-healthcheck == 1.1.1",
   "lua-resty-cookie == 0.1.0",
   "lua-resty-mlcache == 2.4.0",
-  "lua-resty-counter == 0.1.0",
+  "lua-resty-counter == 0.2.0",
   -- external Kong plugins
   "kong-plugin-azure-functions ~> 0.4",
   "kong-plugin-kubernetes-sidecar-injector ~> 0.2",