--- a/spec/01-unit/11-reports_spec.lua
+++ b/spec/01-unit/11-reports_spec.lua
@@ -8,6 +8,7 @@ describe("reports", function()
   describe("send()", function()
     lazy_setup(function()
       reports.toggle(true)
+      reports.create_counter(1)
     end)
 
     lazy_teardown(function()
@@ -82,6 +83,7 @@ describe("reports", function()
       package.loaded["kong.reports"] = nil
       reports = require "kong.reports"
       reports.toggle(true)
+      reports.create_counter(1)
     end)
 
     describe("sends 'database'", function()
@@ -308,6 +310,7 @@ describe("reports", function()
       package.loaded["kong.reports"] = nil
       reports = require "kong.reports"
       reports.toggle(true)
+      reports.create_counter(1)
     end)
 
     it("does not query Redis if not enabled", function()