--- a/spec/fixtures/custom_plugins/kong/plugins/reports-api/api.lua
+++ b/spec/fixtures/custom_plugins/kong/plugins/reports-api/api.lua
@@ -3,6 +3,7 @@ local reports = require "kong.reports"
 return {
   ["/reports/send-ping"] = {
     POST = function()
+      reports._sync_counter()
       reports.send_ping()
       kong.response.exit(200, { message = "ok" })
     end,