### Summary

Previously counters in anonymous reports are implemented with shm
counters. Multiple shm operations happen with request, and each
requires shm level lock across all workers. The lock overhead is
especially noticable while running Kong with large amount of cores.

To improve performance, reports now uses [lua-resty-counter](https://github.com/kong/lua-resty-counter).
It replaces per request shm operation with cheaper worker level Lua
numbers and sync to shm timely.

### Full changelog

* use lock-free lua-resty-counter in reports

### Issues resolved