@fffonion 2019-10-30
RPS comparasion in a 32 workers setup on a 20 cores/40 threads machine

After:
```
root@perf:~# /root/wrk/wrk -t 200 -d 30s -c 200 http://10.88.137.3:8000
Running 30s test @ http://10.88.137.3:8000
  200 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.02ms    3.61ms 127.33ms   96.30%
    Req/Sec   635.47    152.89     1.66k    70.41%
  3808890 requests in 30.10s, 0.85GB read
Requests/sec: 126536.55
Transfer/sec:     28.96MB
root@perf:~# /root/wrk/wrk -t 200 -d 30s -c 200 http://10.88.137.3:8000
Running 30s test @ http://10.88.137.3:8000
  200 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.15ms   27.08ms   1.01s    99.47%
    Req/Sec   638.27    150.77     1.59k    70.46%
  3815010 requests in 30.10s, 0.85GB read
Requests/sec: 126755.48
Transfer/sec:     29.01MB
root@perf:~# /root/wrk/wrk -t 200 -d 30s -c 200 http://10.88.137.3:8000
Running 30s test @ http://10.88.137.3:8000
  200 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.84ms    2.04ms  77.06ms   91.13%
    Req/Sec   646.71    144.96     1.84k    70.32%
  3881118 requests in 30.11s, 0.87GB read
Requests/sec: 128916.47
Transfer/sec:     29.50MB
```

Before
```
root@perf:~# /root/wrk/wrk -t 200 -d 30s -c 200 http://10.88.137.3:8000
Running 30s test @ http://10.88.137.3:8000
  200 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.20ms    2.68ms  82.41ms   94.07%
    Req/Sec   529.06    133.26     1.71k    70.01%
  3171380 requests in 30.10s, 725.72MB read
Requests/sec: 105358.76
Transfer/sec:     24.11MB
root@perf:~# /root/wrk/wrk -t 200 -d 30s -c 200 http://10.88.137.3:8000
Running 30s test @ http://10.88.137.3:8000
  200 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.18ms    2.32ms  84.02ms   91.95%
    Req/Sec   528.07    130.44     1.52k    70.26%
  3168261 requests in 30.10s, 725.01MB read
  Socket errors: connect 0, read 9, write 0, timeout 0
Requests/sec: 105254.96
Transfer/sec:     24.09MB
root@perf:~# /root/wrk/wrk -t 200 -d 30s -c 200 http://10.88.137.3:8000
Running 30s test @ http://10.88.137.3:8000
  200 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.22ms    2.45ms  85.50ms   91.17%
    Req/Sec   515.19    133.61     1.72k    69.41%
  3086729 requests in 30.10s, 706.35MB read
Requests/sec: 102542.47
Transfer/sec:     23.47MB
```
Note: CPU clock locked to 1.2GHz; doesn't reflect maximum RPS Kong can handle on similar hardware.