@fffonion 2019-10-31
I might need a hand on making the CI working,  currently `spec/02-integration/05-proxy/22-reports_spec.lua` is failing because we set lua-resty-counter to sync every 10 minutes. Quickest way is to hardcode a shorter interval in kong/reports.lua to sync counter (like 10ms). But I don't feel it's worthwhile to do that just to make test running.