@fffonion 2019-10-31
@hishamhm I believe we will still be able to send accurate number, just there might be some shifted to next ping window.  Calling`sync` function in lua-resty-counter will only sync current worker's counter but not others if there's any. I can make that change if that's what you are looking for.