@bungle 2019-11-01
Yes, I also think that adding manual (and making possibly the timer optional, e.g. a way to disable it) would be a good addition to library. I guess you might need that to get all the tests green again.