@fffonion 2019-11-01
Thanks for the explanation! That feels a sane way to me to fix tests here. I've opened a PR in library repo and attached both of you as reviewer.
https://github.com/Kong/lua-resty-counter/pull/1