@hishamhm 2019-12-09
@fffonion I think this is good to go now! Could you fix the conflict so we can merge this? Thanks!