@fffonion 2019-12-09
@hishamhm No problem! I merged master into this branch and there's no conflits now.